Immoral hack to make LaunchClassLoader.transformers thread-safe
Normally, only one thread (either the client thread on the dedicated
client, or the server thread on the dedicated server) is actively doing
anything when the game starts up. This is probably why this issue went
unnoticed for so long.

However, our JUnit runner (along with JUnit) is actively loading classes
on a different thread while the game is performing initializtion
(including transformer registrations). This lead to a
ConcurrentModificationException occuring whenever a transformer was
registered at the same time as an off-thread classload.