Create multiple clients without segfaulting Xorg
Re-using the same shared LWJGL shared library won't work, since it holds
a static reference to the GLX context.

Instead, we create a copy of liblwjgl for each new client, and make the
JVM use it by overriding ClassLoader#findLibrary

Note that this is insuffcient for OpenAL, which is loaded manually via
dlopen. We'll probably need to add special handling for it (e.g. copy
some of the LJWL code to locate it)